feat: add CommentItem.helper for comment with user tag
